Add declare winner functionality
Now that the game runs through a full round. I needed a way to make it
stop at some point and say who wins or lose. For now it will be a scor
e of 5 that alerts you that you have lost or not. Next step is to get t
he game to reset the scores.